split connection_strength into 2 files
